Financial RAG Assistant/requirements.txt
faiss-cpu
numpy
tqdm
huggingface-hub
langchain-community
langchain-text-splitters
sentence-transformers
streamlit
tf-keras
python-docx 
docx2txt